Changed testFormula
The testFormula failed due to an Interface that changed
and due to an incorrect value (int instead of double)
return by the ContinuousAtomicProperty.

+++ src/hpnmg/modelchecker/ContinuousAtomicProperty.h
@@ -6,10 +6,10 @@
 
 namespace hpnmg {
     struct ContinuousAtomicProperty {
-        ContinuousAtomicProperty(std::string place, int value) : place(std::move(place)), value(value) {};
+        ContinuousAtomicProperty(std::string place, double value) : place(std::move(place)), value(value) {};
 
         std::string place;
-        int value;
+        double value;
     };
 }
